Make serverless deploy command clean up files afterwards
The 'sls create' command is no longer needed as the script will generate package.json automatically.
The 'sls plugin install ...' is also no longer needed as the script will run this command automatically.
After deployment all 'serverless' files and folders will be removed as it is very messy to keep it around, especially the .serverless folder which for some sad reason contains all the package dependencies (which is also to be found in the Docker image).