[auth] Provide a fallback mechanic for old-style logins
- The identity storage model has changed and all logins should be using
an identity marker. Old users that have been added before this, will not
have the marker and while the idea is to migrate the IDs in the DB we
still need a fallback to where these may have been created during the
upgrade, or where the upgrade for some reason fail to take all the
correct identities into account.